requirements.txt
@'
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
gunicorn==22.0.0
openai>=1.40.0
python-dotenv>=1.0.1
requests>=2.32.3
'@ | Set-Content .\requirements.txt

python -m pip install --upgrade -r .\requirements.txt
python .\master.py

